refactor APOD view in smaller views
